add mot_metrics
