+++ ppdet/metrics/__init__.py
@@ -13,6 +13,8 @@
 # limitations under the License.
 
 from . import metrics
+from . import mot_metrics
 from .metrics import *
+from .mot_metrics import *
 
-__all__ = metrics.__all__
+__all__ = metrics.__all__ + mot_metrics.__all__
